+++ packages/content/pages/GistGitHub.ts
@@ -1,5 +1,6 @@
 import { isGist } from '../utils/PageDetect';
 import { getIconForFile, getIconUrl } from '../utils/Icons';
+import { mutate } from 'fastdom';
 
 const showGistIcons = async () => {
   const fileInfos = document.querySelectorAll('.file-info');
@@ -13,7 +14,9 @@ const showGistIcons = async () => {
     const fileInfo = fileInfos[i] as HTMLDivElement;
     const gistName = (fileInfo.lastElementChild!.firstElementChild as HTMLSpanElement).innerText;
     const iconPath = getIconForFile(gistName);
-    fileInfo.firstElementChild!.innerHTML = `<img src="${getIconUrl(iconPath)}" alt="icon" class="vscode-icon">`;
+    mutate(() => {
+      fileInfo.firstElementChild!.innerHTML = `<img src="${getIconUrl(iconPath)}" alt="icon" class="vscode-icon">`;
+    });
   }
 };