+++ packages/content/pages/SourceForge.ts
@@ -1,5 +1,6 @@
 import { getIconForFile, getIconForFolder, getIconUrl } from '../utils/Icons';
 import { isSourceForgeFiles } from '../utils/PageDetect';
+import { mutate } from 'fastdom';
 
 const QUERY_SOURCEFORGE_ITEMS = '#files_list>tbody>tr';
 
@@ -10,17 +11,16 @@ function showIconsForFiles() {
 
     const newIconEl = document.createElement('img');
     newIconEl.setAttribute('class', 'vscode-icon sf-icon');
-    const iconAndNameEl = item.firstElementChild as HTMLTableHeaderCellElement;
+    const iconAndNameEl = item.firstElementChild.firstElementChild as HTMLTableHeaderCellElement;
 
     const isFolder = item.className.includes('folder');
     if (isFolder) {
       /**
        * [TR:
-       *  [TH: [SPAN: icon]],
+       *  [TH: [A: [SVG: icon], [SPAN: folderName]]],
        *  [TD: [ABBR: date]],
        *  [TD: size],
-       *  [TD: [A: [IMG: chart]]],
-       *  [TD: [SPAN: info]]
+       *  [TD: [DIV: Populated by JS], [DIV: [A: chart]]],
        * ]
        */
 
@@ -28,24 +28,27 @@ function showIconsForFiles() {
       const nameEl = iconAndNameEl.lastElementChild as HTMLAnchorElement;
       const name = nameEl.innerText.toLowerCase();
       const iconPath = getIconForFolder(name);
-      newIconEl.setAttribute('src', getIconUrl(iconPath));
-      iconAndNameEl.replaceChild(newIconEl, iconEl);
+      mutate(() => {
+        newIconEl.setAttribute('src', getIconUrl(iconPath));
+        iconAndNameEl.replaceChild(newIconEl, iconEl);
+      });
     } else {
       /**
        * [TR:
-       *  [TH: [[SPAN: icon], [A: name]]],
+       *  [TH: [A: [SVG: icon]]],
        *  [TD: [ABBR: date]],
        *  [TD: size],
-       *  [TD: [A: [IMG: chart]]],
-       *  [TD: [SPAN: info]]
+       *  [TD: [DIV: Populated by JS], [DIV: [A: chart]]],
        * ]
        */
 
       const nameEl = iconAndNameEl.firstElementChild as HTMLAnchorElement;
       const name = nameEl.innerText.toLowerCase();
       const iconPath = getIconForFile(name);
-      newIconEl.setAttribute('src', getIconUrl(iconPath));
-      iconAndNameEl.insertBefore(newIconEl, nameEl);
+      mutate(() => {
+        newIconEl.setAttribute('src', getIconUrl(iconPath));
+        iconAndNameEl.insertBefore(newIconEl, nameEl);
+      });
     }
   }
 }