+++ typings/fastdom/index.d.ts
@@ -1,0 +1,5 @@
+declare module "fastdom" {
+    export const clear: () => void;
+    export const measure: (callback: () => void) => void;
+    export const mutate: (callback: () => void) => void;
+}