+++ typings/languages-vscode.d.ts
@@ -1,0 +1,9 @@
+declare module "*languages-vscode.json" {
+    const json: {
+        [languageId: string]: {
+            extensions: string[];
+            filenames: string[];
+        }
+    };
+    export default json;
+}