+++ typings/languages-vsi.d.ts
@@ -1,0 +1,8 @@
+declare module "*languages-vsi.json" {
+    const json: {
+        [languageId: string]: {
+            defaultExtension: string;
+        }
+    };
+    export default json;
+}